plot signal attempt5.

+++ tradevisor-app/src/main/java/ru/grnk/tradevisor/calculate/strategies/dto/Marker.java
@@ -1,4 +1,4 @@
 package ru.grnk.tradevisor.calculate.strategies.dto;
 
-public record Marker(int x, double y, String color) {
+public record Marker(int x, Float y, String color) {
 }
