+++ tradevisor-app/src/main/java/ru/grnk/tradevisor/calculate/strategies/dto/RenderData.java
@@ -4,13 +4,13 @@ import java.util.List;
 public record RenderData(
         List<OhlcRecord> data,
         List<Integer> fiboXaxe,
-        double fibo382,
-        double fibo618,
+        Float fibo382,
+        Float fibo618,
         List<Marker> markers,
-        Double priceOpen,
-        double stopLoss,
-        double takeProfit,
-        double infimum,
-        double supremum
+        Float priceOpen,
+        Float stopLoss,
+        Float takeProfit,
+        Float infimum,
+        Float supremum
 ) {
 }
